Manual revisions
Mainly reorganizing test files.
