+++ tests/test_subplots.py
@@ -1,20 +1,14 @@
 """Tests for the `subplots` module and subplot integration with GUIs."""
 
-from collections import abc
-
-import ipywidgets as w
 import pandas as pd
 import pytest
 
-from market_analy import analysis, charts
-from market_analy.formatters import formatter_datetime
 from market_analy.subplots import (
     SUBPLOT_REGISTRY,
     Subplot,
     _volume,
     resolve_subplots,
 )
-from market_analy.utils.ipywidgets_utils import capture_widgets
 
 
 def _close(prices: pd.DataFrame) -> pd.Series:
@@ -69,175 +63,3 @@ class TestVolumeDataCreator:
         prices = pd.DataFrame([[1.0], [2.0], [3.0]], index=idx, columns=cols)
         with pytest.raises(ValueError, match="does not include a 'volume' column"):
             _volume(prices)
-
-
-class TestGuiSubplots:
-    """Tests for subplots stacked beneath a price chart GUI."""
-
-    @pytest.fixture
-    def analy(self, prices_analysis) -> abc.Iterator[analysis.Analysis]:
-        yield analysis.Analysis(prices_analysis)
-
-    @pytest.fixture
-    def pp(self) -> dict:
-        return {"start": pd.Timestamp("2023-01-06"), "end": pd.Timestamp("2023-01-10")}
-
-    def test_no_subplots_default(self, analy, pp):
-        """Without `subplots` the gui has no subplots."""
-        gui = analy.plot(**pp, display=False)
-        assert gui._subplots == []
-        assert gui.chart.figure in gui._gui_box_contents
-        # x-tick labels remain visible on the price chart
-        assert "display" not in gui.chart.axes[0].tick_style
-
-    def test_subplots_shared_scale(self, analy, pp):
-        """Subplot reuses the price chart's x-axis scale."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        assert len(gui._subplots) == 1
-        pane = gui._subplots[0]
-        assert isinstance(pane, charts.SubplotBars)
-        assert pane.scales["x"] is gui.chart.scales["x"]
-        assert pane.x_ticks.equals(gui.chart.x_ticks)
-
-    def test_subplots_in_gui_box(self, analy, pp):
-        """Subplot figure sits between the price chart and the date slider."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        contents = gui._gui_box_contents
-        i_chart = contents.index(gui.chart.figure)
-        i_pane = contents.index(gui._subplots[0].figure)
-        i_slider = contents.index(gui.date_slider)
-        assert i_chart < i_pane < i_slider
-
-    # TODO: still to review tests from here...
-    def test_subplots_data(self, analy, pp):
-        """Subplot data reflects the data_creator output."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        expected = _volume(gui.prices)
-        assert (pane.data.values == expected.values).all()
-
-    def test_subplots_multiple_ordered(self, analy, pp):
-        """Multiple subplots stack, in order, all sharing the x scale."""
-        custom = Subplot(data_creator=_close, kind="lines", title="Close")
-        gui = analy.plot(**pp, subplots=["volume", custom], display=False)
-        assert len(gui._subplots) == 2
-        assert isinstance(gui._subplots[0], charts.SubplotBars)
-        assert isinstance(gui._subplots[1], charts.SubplotLines)
-        scale = gui.chart.scales["x"]
-        assert all(p.scales["x"] is scale for p in gui._subplots)
-        contents = gui._gui_box_contents
-        assert contents.index(gui._subplots[0].figure) < contents.index(
-            gui._subplots[1].figure
-        )
-
-    def test_x_label_policy(self, analy, pp):
-        """X-tick labels show only on the bottom-most pane."""
-        custom = Subplot(data_creator=_close, kind="lines", title="Close")
-        gui = analy.plot(**pp, subplots=["volume", custom], display=False)
-        assert gui.chart.axes[0].tick_style.get("display") == "none"
-        assert gui._subplots[0].axes[0].tick_style.get("display") == "none"
-        assert "display" not in (gui._subplots[-1].axes[0].tick_style or {})
-
-    def test_shared_scale_lockstep(self, analy, pp):
-        """Changing the price chart's plotted window moves the subplot too."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        x_ticks = gui.chart.x_ticks
-        narrow = pd.Interval(x_ticks[1], x_ticks[-1], closed="left")
-        gui.chart.plotted_x_ticks = narrow
-        assert pane.plotted_x_ticks.equals(gui.chart.plotted_x_ticks)
-
-    def test_update_subplots(self, analy, pp):
-        """Recomputing subplots from new prices refreshes the pane data."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        new_prices = gui.prices * 2
-        gui._update_subplots(new_prices)
-        expected = _volume(new_prices)
-        assert (pane.data.values == expected.values).all()
-
-    def test_update_chart_funnel_refreshes_subplots(self, analy, pp):
-        """The `_update_chart` funnel refreshes subplots when prices change."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        new_prices = gui.prices * 2
-        data, data2 = gui._prices_to_chart_data(new_prices)
-        gui._update_chart(data, data2, prices=new_prices)
-        expected = _volume(new_prices)
-        assert (pane.data.values == expected.values).all()
-
-    def test_reset_chart_keeps_subplots(self, analy, pp):
-        """Resetting the chart leaves the subplot data intact."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        gui._reset_chart()
-        expected = _volume(gui.prices)
-        assert (pane.data.values == expected.values).all()
-
-    def test_ref_levels_add_marks(self, analy, pp):
-        """A subplot with `ref_levels` adds persistent reference marks."""
-        custom = Subplot(
-            data_creator=_close, kind="lines", title="Close", ref_levels=[1.0]
-        )
-        gui = analy.plot(**pp, subplots=[custom], display=False)
-        pane = gui._subplots[0]
-        assert len(pane.added_marks[charts.Groups.PERSIST]) == 1
-
-    def test_widget_capture_restored_after_subplots(self, analy, pp):
-        """GUI widgets built after the panes are still tracked in _widgets."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        # a widget created after the panes is tracked against the gui
-        assert gui.date_slider in gui._widgets
-        # the pane's own widgets live on the pane, not the gui
-        assert pane.figure in pane._widgets
-        assert pane.figure not in gui._widgets
-
-    def test_subplot_title_above_plot(self, analy, pp):
-        """The subplot title is the figure title, with room above the plot."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        assert pane.figure.title == "Volume"
-        # the title is no longer rendered as the y-axis label
-        assert pane.axes[1].label in (None, "")
-        # top margin reserves room so the title sits above the plot area
-        assert pane.figure.fig_margin["top"] >= 20
-
-    def test_volume_subplot_tooltip(self, analy, pp):
-        """Hovering a volume bar yields a tooltip with the bar's value."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        pane = gui._subplots[0]
-        assert isinstance(pane, charts.SubplotBars)
-        assert pane._has_tooltip
-        assert pane.mark.tooltip is not None
-        y = float(pane.data.iloc[0])
-        html = pane._tooltip_value(pane.mark, {"data": {"index": 0, "y": y}})
-        assert "Volume" in html
-        assert f"{int(y):,}" in html
-        assert formatter_datetime(pane.x_ticks[0]) in html
-
-    def test_close(self, analy, pp):
-        """Closing the gui closes subplots without error."""
-        gui = analy.plot(**pp, subplots=["volume"], display=False)
-        gui.close()
-
-
-class TestCaptureWidgets:
-    """Tests for the `capture_widgets` context manager."""
-
-    def test_capture_and_restore(self):
-        initial = w.Widget._widget_construction_callback
-        outer: list[w.Widget] = []
-        inner: list[w.Widget] = []
-        with capture_widgets(outer):
-            a = w.HTML()
-            with capture_widgets(inner):
-                b = w.HTML()
-            c = w.HTML()
-        assert a in outer
-        # the outer capture resumes once the inner context exits
-        assert c in outer
-        assert b in inner
-        assert b not in outer
-        # the prior callback is restored (not simply cleared)
-        assert w.Widget._widget_construction_callback is initial
