+++ tests/tests_utils/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for modules of the `utils` subpackage."""
